Moved the declarations of missing S3LevelDown methods to /types/
